services/subscriptions/func-subscriptions-automation/requirements.txt
azure-functions
azure-cli
azure-identity
azure-data-tables
azure-storage-blob
azure-mgmt-monitor
azure-mgmt-consumption
azure-mgmt-resource 
azure-mgmt-storage
azure-keyvault-secrets
pytest
pytest_mock
black
python-dateutil
pytest-cov
python-dotenv
pytz
requests
openpyxl